Split up lib.rs
